requirements.txt
groq
python-dotenv
rich
fastapi
uvicorn
pydantic